requirements.txt
openai-whisper>=20231117
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.20.0
